fix: review round - stop blocking a pool thread, share the handler body
Qodo found the one that matters. After the receive became awaited, the
continuation runs on a thread-pool thread, and SchedulerMessageHandler's
finally blocks it waiting for scheduler capacity. Before this branch that
block landed on the worker's own dedicated thread, where blocking is free, so
this was a regression introduced here - and blocking a pool thread is the
exact starvation the branch exists to remove.

MessageProcessingAsync now awaits that capacity itself, so the block further
down finds room and returns immediately. It goes through ATaskScheduler rather
than resolving the wait handle directly: for a scheduler consumer the user's
scheduler is registered into the consumer container from a different container,
so the directly resolved singleton would have been a different instance and the
gate would have been meaningless. The result is deliberately ignored - on
cancellation, falling through lets ShouldHandle throw as it always has, so the
message is rolled back rather than silently dropped.

The two Redis query handlers now share a base class holding everything except
the call that runs the Lua script. They were near-identical on purpose, on the
argument that a difference between two copies would surface on only one code
path; one implementation is a stronger guarantee of the same property, and it
means the synchronous path's existing tests cover the asynchronous one. It also
clears Sonar's duplication gate and most of the patch-coverage gap, which were
the same problem seen twice.

Also from the round: forward the cancellation token into the Polly pipeline, or
a retry keeps waiting out its delay after the queue is told to stop, which was
the reliability rating Sonar failed on. Honour the caller's token at both loop
checks in the Redis receive and link it into the work-signal wait - it was
accepted and ignored, which is worse than not taking it. Guard the pub/sub
callback against a handle disposed underneath it. Reconcile the control test,
whose documented measurements contradicted its own assertion.

Not fixed here and tracked in #284: the synchronous Remove on the expiry
branch, flagged by both reviewers. It needs a new member on IRemoveMessage, the
same breaking change as the commit, rollback and history calls in the same
position, and those are better done as one release than piecemeal.
